refactor: extract utils to prevent circular dependencies
