+++ apps/server/src/controllers/integrationController.config.ts
@@ -1,5 +1,6 @@
 import { OntimeEvent, isKeyOfType, isOntimeEvent } from 'ontime-types';
-import { editEvent, getEventWithId } from '../services/rundown-service/RundownService.js';
+import { editEvent } from '../services/rundown-service/RundownService.js';
+import { getEventWithId } from '../services/rundown-service/rundownUtils.js';
 import { coerceString, coerceNumber, coerceBoolean, coerceColour } from '../utils/coerceType.js';
 
 const whitelistedPayload = {
