+++ apps/server/src/controllers/rundownController.ts
@@ -10,11 +10,10 @@ import {
   deleteAllEvents,
   deleteEvent,
   editEvent,
-  getRundown,
   reorderEvent,
   swapEvents,
 } from '../services/rundown-service/RundownService.js';
-import { get as getCachedRundown } from '../services/rundown-service/rundownCache.js';
+import { getNormalisedRundown, getRundown } from '../services/rundown-service/rundownUtils.js';
 
 // Create controller for GET request to '/events'
 // Returns -
@@ -26,7 +25,7 @@ export const rundownGetAll: RequestHandler = async (_req, res) => {
 // Create controller for GET request to '/events/cached'
 // Returns -
 export const rundownGetCached: RequestHandler = async (_req: Request, res: Response<RundownCached>) => {
-  const cachedRundown = getCachedRundown();
+  const cachedRundown = getNormalisedRundown();
   res.json(cachedRundown);
 };
